Redirect to the login page on a authentication failure
This is NOT ideal, forcing a redirect is bad UX and the user may
lose work.

Instead, we should display on error in the page where the user is
and not redirect.

This will be refactored to remove the Axios interceptor and use
React hooks for fetching the data and maintaining any error state.

A common error component can also be provided to prompt the user
to login again.